requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
motor>=3.4.0
python-dotenv
